requirements.txt
langchain
openai
python-dotenv
google-search-results
numexpr
